backend/requirements.txt
# ─── FastAPI ─────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ─── Database ────────────────────────────
motor==3.4.0
pymongo==4.7.2
redis==5.0.4

# ─── Auth ───────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ─── PDF ────────────────────────────────
pymupdf==1.24.4

# ─── AI / ML ────────────────────────────
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4

# ─── Whisper (FIXED) ────────────────────
faster-whisper==1.0.3
imageio-ffmpeg==0.4.9

# ─── LLM ────────────────────────────────
groq==0.8.0

# ─── Core Utils ─────────────────────────
aiofiles==23.2.1
httpx==0.27.0
python-dotenv==1.0.1

# ─── Pydantic ───────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1

# ─── Testing (optional but fine) ─────────
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0